ADD|UPDATE - service API untuk operasi CRUD peserta (insert, update, delete, get) menggunakan Retrofit; ubah nama Participants.kt

+++ app/src/main/java/com/example/eventorganizer/model/Participants.kt
@@ -20,7 +20,7 @@ data class ParticipantDetailResponse (
 @Serializable
 data class Participant (
     @SerialName("id_peserta")
-    val idPeserta: Long,
+    val idPeserta: Int,
 
     @SerialName("nama_peserta")
     val namaPeserta: String,
